+++ src/kolesov93/RunnerPlayer/SongProvider.java
@@ -4,5 +4,7 @@ package kolesov93.RunnerPlayer;
  * Created by kolesov93
  */
 public interface SongProvider {
-    public Song getSongForSpeed(int speed);
+    public Song getSongForSpeed(double speed);
 }
+
+
